containerService/containerService/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
docker==6.1.3
grpcio==1.58.0
grpcio-tools==1.58.0
idna==3.4
packaging==23.1
protobuf==4.24.3
requests==2.31.0
urllib3==2.0.4
websocket-client==1.6.3
